requirements.txt
requests
python-dotenv
pyyaml
plotly
pandas
flask